requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.18.0
motor>=3.1.1
pydantic>=2.0.0
pydantic-settings>=2.0.0   # <-- add this
PyJWT>=2.8.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
apscheduler>=3.11.0
pandas>=2.0.0
numpy>=1.24.0
python-json-logger>=2.0.4
pytest>=7.2.0
pytest-asyncio>=0.22.0
httpx>=0.24.0
